cleanup, refactoring, reorganization
