+++ all/src/index.ts
@@ -1,4 +1,4 @@
-import merge from 'deepmerge';
+import * as merge from 'deepmerge';
 
 import { typography } from '~/typography';
 import { colors } from '~/colors';
