+++ colors/doc/index.mdx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import { mdx } from '@mdx-js/react';
 import { TailwindShowcases } from '@divriots/dockit-react/tailwind-showcases';
-import { theme } from '~/tailwind';
+import { theme } from '~/all';
 import { DefaultLayout } from '~/mdx-layout';
 export default DefaultLayout;
