+++ disclosure/doc/disclosure.mdx
@@ -12,7 +12,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { Disclosure } from '~/disclosure';
+import { Disclosure } from '@divriots/starter-tailwindcss/disclosure';
 
 <Disclosure
   title="What is your refund policy?"
