+++ listbox/doc/listbox.mdx
@@ -12,7 +12,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { Listbox } from '~/listbox';
+import { Listbox } from '@divriots/starter-tailwindcss/listbox';
 
 <Listbox
   items={[
