+++ mdx-layout/src/index.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import * as React from 'react';
 import { MDXProvider } from '@mdx-js/react';
 import { CoreLayout } from '@divriots/dockit-react/mdx-layout-core/dist/CoreLayout';
-import '~/tailwind/src/styles.css';
+import '~/all/src/styles.css';
 
 export const DefaultLayout = (props) => (
   <MDXProvider components={{}}>
