+++ menu/doc/menu.mdx
@@ -13,7 +13,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { Menu } from '~/menu';
+import { Menu } from '@divriots/starter-tailwindcss/menu';
 import { PencilIcon, DuplicateIcon } from '@heroicons/react/solid';
 
 <Menu
