+++ menu/src/menu.tsx
@@ -1,4 +1,5 @@
-import React, { Fragment, ReactNode } from 'react';
+import * as React from 'react';
+import { Fragment, ReactNode } from 'react';
 import { Menu as HMenu, Transition } from '@headlessui/react';
 import { ChevronDownIcon } from '@heroicons/react/solid';
 
@@ -24,20 +25,18 @@ type MenuProps = {
   items?: MenuItemDetails[];
 };
 
-const MenuItem = ({ Icon, label, onClick }: MenuItemDetails & { key: any }) => (
+const MenuItem = ({ Icon, label, onClick }: MenuItemDetails & { key: any; }) => (
   <HMenu.Item>
     {({ active }) => (
       <button
-        className={`${
-          active ? 'bg-purple-500 text-white' : 'text-grey-900'
-        } group flex rounded-md items-center w-full px-2 py-2 text-sm`}
+        className={`${active ? 'bg-purple-500 text-white' : 'text-grey-900'
+          } group flex rounded-md items-center w-full px-2 py-2 text-sm`}
         onClick={onClick}
       >
         {Icon && (
           <Icon
-            className={`w-5 h-5 mr-2 ${
-              active ? 'text-white' : 'text-grey-900'
-            }`}
+            className={`w-5 h-5 mr-2 ${active ? 'text-white' : 'text-grey-900'
+              }`}
           />
         )}
         {label}
