+++ popover/doc/popover.mdx
@@ -17,7 +17,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { Popover } from '~/popover';
+import { Popover } from '@divriots/starter-tailwindcss/popover';
 
 <Popover
   title="Solutions"
