+++ radio-group/doc/radio-group.mdx
@@ -12,7 +12,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { RadioGroup } from '~/radio-group';
+import { RadioGroup } from '@divriots/starter-tailwindcss/radio-group';
 
 <RadioGroup
   label="Server Size"
