+++ readme/doc/index.mdx
@@ -1,6 +1,8 @@
-import React from 'react';
+import * as React from 'react';
 import { mdx } from '@mdx-js/react';
 import { DefaultLayout } from '~/mdx-layout';
 export default DefaultLayout;
 
 # TailwindCSS Design System
+
+### TODO
