+++ shadows/doc/index.mdx
@@ -1,6 +1,6 @@
 import { mdx } from '@mdx-js/react';
 import { TailwindShowcases } from '@divriots/dockit-react/tailwind-showcases';
-import { theme } from '~/tailwind';
+import { theme } from '~/all';
 import { DefaultLayout } from '~/mdx-layout';
 export default DefaultLayout;
