+++ switch/doc/switch.mdx
@@ -12,7 +12,7 @@ export default DefaultLayout;
 ## Usage
 
 ```tsx
-import { Switch } from '~/switch';
+import { Switch } from '@divriots/starter-tailwindcss/switch';
 
 <Switch onSwitched={(enabled) => console.log('Switched to', enabled)} />;
 ```
