+++ switch/index.ts
