+++ switch/src/switch.tsx
@@ -1,4 +1,5 @@
-import React, { useState } from 'react';
+import * as React from 'react';
+import { useState } from 'react';
 import { Switch as HSwitch } from '@headlessui/react';
 
 export type SwitchProps = {
@@ -12,7 +13,7 @@ export type SwitchProps = {
  Switches are a pleasant interface for toggling a value between two states,
  and offer the same semantics and keyboard navigation as native checkbox elements.
 */
-export const Switch = ({ onSwitched = () => {} }: SwitchProps) => {
+export const Switch = ({ onSwitched = () => { } }: SwitchProps) => {
   const [enabled, setEnabled] = useState(false);
 
   return (
