+++ tsconfig.json
@@ -8,10 +8,9 @@
     "reactNamespace": "React",
     "module": "ESNext",
     "moduleResolution": "Node",
+    "baseUrl": ".",
     "paths": {
-      "~/*": [
-        "/*"
-      ]
+      "~/*": ["*"]
     }
   }
 }
